[CHANGE] Updates on tsne file and main WE module. fix internal paths
